% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_arrets_tan.R
\docType{data}
\name{arrets_tan}
\alias{arrets_tan}
\title{Arrets de la compagnie de transport nantaise tan.}
\format{
\subsection{\code{arrets_tan}}{

Data frame 2838 lignes x 3 colonnes
\describe{
\item{codeLieu}{Code de référence de l'arrêt}
\item{libelle}{Nom de l'arret complet}
\item{ligne}{nom ou numéro de ligne}
}
}
}
\source{
\url{https://open.tan.fr/ewp/arrets.json}
}
\usage{
arrets_tan
}
\description{
Ce jeux de données comporte la liste des arrêts de la tan, avec leur nom
leur libéllé, et les numéros ou noms de ligne de transport qui y passent.
Les données ont été récoltées le 29-01-2023.
}
\keyword{datasets}
