% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_name.R
\name{find_closest_name}
\alias{find_closest_name}
\title{find_closet_name}
\usage{
find_closest_name(list_name, name)
}
\arguments{
\item{list_name}{Liste de nom qui sera utilisée et confrontée à name}

\item{name}{Nom que l'on souhaite trouver dans la liste}
}
\value{
Elle retourne une vecteur de chaine de cacartère prenant ayant une
longueur de 1 à n où n est la taille de la liste donnée en entrée.
}
\description{
Cette fonction prend en entré une chaine de caractères et une liste
de chaines de caractères. Elle renvoie le ou les chaines de caractères
de la liste qui ont l'écriture la plus proche de la chaine caractère
seule. Ainsi elle peut-être utilisé dans le cadre de correction de saisie.
Pour trouver les mots qui se rapprochent le plus nous utlisons la
méthode de Levenshtein. Elle ne différencie pas les majuscules des
miniscules. Et ne renvoie qu'une fois une valeur répétée.
}
\examples{
find_closest_name(list_name = arrets_tan$libelle,name = "Morr")
find_closest_name(list_name = arrets_tan$libelle,name = "Commerce")
}
