% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_transport.R
\name{get_next_transport}
\alias{get_next_transport}
\title{get_next_transport}
\usage{
get_next_transport(station)
}
\arguments{
\item{station}{Le paramètre station est une chaine de caractères.}
}
\value{
La fonction renvoie un data frame comportant le nom de la ligne,
son terminus et le temps d'attente pour l'arrêt spécifié.
}
\description{
Cette fonction permet de renvoyer les horaires de passages des lignes
de transport du réseau nantais. Il faut pour cela donner le nom
d'une station en entrée. Si le nom est incorect, la fonction renvoie
la où les stations qui ont le nom le plus proche au sens de Levenshtein.
}
\examples{
get_next_transport("Commerce")
}
