% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_transport_addin.R
\name{get_next_transport_addin}
\alias{get_next_transport_addin}
\title{get_next_transport_addin}
\usage{
get_next_transport_addin()
}
\value{
La sortie de la fonction get_next_transport
}
\description{
La fonction appelle la fonction get_next_transport via un addin.
La fonction ne prend pas d'argument et renvoie l'utilisateur sur une interface shiny.
Il est possible de rentrer le nom de l'arret et de cliquer sur un bouton,
pour afficher les résultats. Un second bouton est disponible pour se déloguer de la session
locale.
}
