% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waiting_time.R
\name{get_waiting_time}
\alias{get_waiting_time}
\title{get_waiting_time : obtenir le temps d'attente.}
\source{
\url{https://open.tan.fr/ewp/tempsattente.json/CODELIEU}
}
\usage{
get_waiting_time(code_lieu)
}
\arguments{
\item{code_lieu}{Le paramétre code lieu prend en compte le code de l'arrêt.}
}
\value{
La fonction affiche un dataframe de trois colonnes: numéro de la ligne, terminus et temps d'attente.
}
\description{
Cette fonction prend en entrée le code_lieu d'un arrêt et renvoie le temps d'attente
des différentes lignes de transports à cet arrêt. Le terminus des lignes est précisé
pour nous indiquer dans quel sens va la ligne.
}
\examples{
get_waiting_time("AMER") #renvoie les temps d'attentes à l'arrêt américain.
}
