% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotan-package.R
\docType{package}
\name{infotan-package}
\alias{infotan}
\alias{infotan-package}
\title{infotan: Données en temps réel des transports nantais}
\description{
Le package récupère les données de la tan en temps réel pour permettre aux utilisateurs d'appeller la fonction get_waiting time qui leur permet d'obtenir ces information en tapant le nom d'un arrêt. Dans cette librairie vous trouverez une base de données nommée arrêts_tan qui retranscrit la liste des arrêts de la tan.
}
\author{
\strong{Maintainer}: Benoit-Alexandre Tranvouez \email{benoit-alexandre.tranvouez@etu.univ-nantes.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
