% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_station.R
\name{is_station}
\alias{is_station}
\title{La fonction is_station}
\usage{
is_station(station)
}
\arguments{
\item{station}{Le nom de la station en toutes lettres}
}
\value{
Elle renvoie un booléen qui permet de déterminer si la chaine
appartient "TRUE" ou non "FALSE" à la liste d'arrêt de la tan.
La fonction gére le problème de majuscule.
}
\description{
Cette fonction prend en entrée une chaine de caractère dans l'optique de
déterminer si elle correspond au nom d'un arrêt du réseau nantais.
}
\examples{
is_station("Bruneau")#TRUE
is_station("Tennis")#FALSE
is_station("bruneau")#TRUE
}
